requirements.txt
asgiref==3.6.0
Django==4.1.7
django-cors-headers==3.14.0
djangorestframework==3.14.0
mysqlclient==2.1.1
Pillow==9.4.0
psycopg2==2.9.5
python-decouple==3.8
asgiref==3.6.0
certifi==2022.12.7
distlib==0.3.6
Django==4.1.7
django-cors-headers==3.14.0
djangorestframework==3.14.0
filelock==3.9.0
Pillow==9.4.0
pipenv==2023.2.18
platformdirs==3.1.1
psycopg2==2.9.5
python-decouple==3.8
python-dotenv==1.0.0
pytz==2022.7.1
sqlparse==0.4.3
tzdata==2022.7
virtualenv==20.20.0
virtualenv-clone==0.5.7